move to a virtual workspace
