+++ rta-for-fps-lib/src/iterators.rs
