+++ rta-for-fps-lib/src/iterators/curve.rs
