+++ rta-for-fps-lib/src/iterators/curve/aggregate.rs
