+++ rta-for-fps-lib/src/iterators/curve/delta.rs
