+++ rta-for-fps-lib/src/iterators/curve/split.rs
