+++ rta-for-fps-lib/src/iterators/peek.rs
