+++ rta-for-fps-lib/tests/consolidated_tests/server_tests.rs
@@ -1,9 +1,9 @@
-use rta_for_fps::curve::Curve;
-use rta_for_fps::iterators::CurveIterator;
-use rta_for_fps::server::{Server, ServerKind};
-use rta_for_fps::task::Task;
-use rta_for_fps::time::TimeUnit;
-use rta_for_fps::window::Window;
+use crate::rta_lib::curve::Curve;
+use crate::rta_lib::iterators::CurveIterator;
+use crate::rta_lib::server::{Server, ServerKind};
+use crate::rta_lib::task::Task;
+use crate::rta_lib::time::TimeUnit;
+use crate::rta_lib::window::Window;
 
 #[test]
 fn deferrable_server() {
